Add role deletion functionality across API and GraphQL
A new endpoint for deleting a role has been added to the `RolesController` in the `DynSec.API.Controllers.DynSec` namespace. This endpoint is defined as a POST request to `api/<MQTTdynsecController>/role/{role}/delete` and handles the deletion of a role by calling the `DeleteRole` method from the `rolesService`. It returns a 504 status code if a `DynSecProtocolInvalidParameterException` is caught and a 404 status code if a `DynSecProtocolNotFoundException` is caught.

In the `DynSec.GraphQL` namespace, a new method `DeleteRoleAsync` has been added to the `DynSecMutation` class. This method calls the `DeleteRole` method from the `rolesService` to delete a role asynchronously.

The `IRolesService` interface in the `DynSec.Protocol.Interfaces` namespace has been updated to include a new method `DeleteRole` which takes a string parameter representing the role to be deleted.

In the `RolesServiceMutations` class within the `DynSec.Protocol` namespace, a new method `DeleteRole` has been implemented. This method checks if the role parameter is null and throws a `DynSecProtocolInvalidParameterException` if it is. It then creates a `DeleteRole` command and executes it, returning a confirmation string upon successful execution.