Refactor GroupDetailComponent for client management
Updated the HTML template to replace the roles priority list with a clients priority list. Added `clientsChanged` and `selectedClients` properties in the TypeScript file. Modified the data source for `allClients` to use `userName` and updated the `updateSelectedItems` method to initialize `selectedClients`. Changed the `clients` property in the `Group` interface to use an array of `Client` objects for better structure.